src/test/javascript/spec/app/entities/competitionservicejoined/competitionservicejoined.service.spec.ts
import { TestBed, getTestBed } from '@angular/core/testing';
import { HttpClientTestingModule, HttpTestingController } from '@angular/common/http/testing';
import { take, map } from 'rxjs/operators';
import { CompetitionservicejoinedService } from 'app/entities/competitionservicejoined/competitionservicejoined.service';
import { ICompetitionservicejoined, Competitionservicejoined } from 'app/shared/model/competitionservicejoined.model';
import { ServicesState } from 'app/shared/model/enumerations/services-state.model';

describe('Service Tests', () => {
  describe('Competitionservicejoined Service', () => {
    let injector: TestBed;
    let service: CompetitionservicejoinedService;
    let httpMock: HttpTestingController;
    let elemDefault: ICompetitionservicejoined;
    let expectedResult: ICompetitionservicejoined | ICompetitionservicejoined[] | boolean | null;
    beforeEach(() => {
      TestBed.configureTestingModule({
        imports: [HttpClientTestingModule]
      });
      expectedResult = null;
      injector = getTestBed();
      service = injector.get(CompetitionservicejoinedService);
      httpMock = injector.get(HttpTestingController);

      elemDefault = new Competitionservicejoined(0, ServicesState.ACCEPT, 'AAAAAAA');
    });

    describe('Service methods', () => {
      it('should find an element', () => {
        const returnedFromService = Object.assign({}, elemDefault);
        service
          .find(123)
          .pipe(take(1))
          .subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'GET' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject(elemDefault);
      });

      it('should create a Competitionservicejoined', () => {
        const returnedFromService = Object.assign(
          {
            id: 0
          },
          elemDefault
        );
        const expected = Object.assign({}, returnedFromService);
        service
          .create(new Competitionservicejoined())
          .pipe(take(1))
          .subscribe(resp => (expectedResult = resp.body));
        const req = httpMock.expectOne({ method: 'POST' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject(expected);
      });

      it('should update a Competitionservicejoined', () => {
        const returnedFromService = Object.assign(
          {
            state: 'BBBBBB',
            details: 'BBBBBB'
          },
          elemDefault
        );

        const expected = Object.assign({}, returnedFromService);
        service
          .update(expected)
          .pipe(take(1))
          .subscribe(resp => (expectedResult = resp.body));
        const req = httpMock.expectOne({ method: 'PUT' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject(expected);
      });

      it('should return a list of Competitionservicejoined', () => {
        const returnedFromService = Object.assign(
          {
            state: 'BBBBBB',
            details: 'BBBBBB'
          },
          elemDefault
        );
        const expected = Object.assign({}, returnedFromService);
        service
          .query()
          .pipe(
            take(1),
            map(resp => resp.body)
          )
          .subscribe(body => (expectedResult = body));
        const req = httpMock.expectOne({ method: 'GET' });
        req.flush([returnedFromService]);
        httpMock.verify();
        expect(expectedResult).toContainEqual(expected);
      });

      it('should delete a Competitionservicejoined', () => {
        service.delete(123).subscribe(resp => (expectedResult = resp.ok));

        const req = httpMock.expectOne({ method: 'DELETE' });
        req.flush({ status: 200 });
        expect(expectedResult);
      });
    });

    afterEach(() => {
      httpMock.verify();
    });
  });
});
